requirements.txt
streamlit
streamlit-option-menu
streamlit-lottie
streamlit-extras
streamlit-custom-notification-box
python-docx
pandas
plotly
pillow
pycryptodome==3.20.0
python-dotenv
pdfminer.six==20221105
nltk
scikit-learn
sqlalchemy
openpyxl
requests
spacy
pypdf==4.2.0
selenium
numpy
webdriver-manager
chromedriver-autoinstaller
google-generativeai
pdf2image
pytesseract
pdfplumber
reportlab
openrouter
docx2pdf
docx2txt
python-pptx
matplotlib
seaborn
pypdf2
